+++ packages/bundle/src/adaptiveCards/Attachment/AnimationCardAttachment.js
@@ -1,29 +1,14 @@
-/* eslint react/no-array-index-key: "off" */
-
-import { Components, hooks } from 'botframework-webchat-component';
 import PropTypes from 'prop-types';
 import React from 'react';
 
-import CommonCard from './CommonCard';
+import AnimationCardContent from './AnimationCardContent';
 
-const { ImageContent, VideoContent } = Components;
-const { useStyleSet } = hooks;
+const AnimationCardAttachment = ({ attachment: { content }, disabled }) => (
+  <AnimationCardContent content={content} disabled={disabled} />
+);
 
-const AnimationCardAttachment = ({ attachment, attachment: { content: { media = [] } } = {} }) => {
-  const [{ animationCardAttachment: animationCardAttachmentStyleSet }] = useStyleSet();
-
-  return (
-    <div className={animationCardAttachmentStyleSet}>
-      <ul className="media-list">
-        {media.map(({ profile = '', url }, index) => (
-          <li key={index}>
-            {/\.gif$/iu.test(url) ? <ImageContent alt={profile} src={url} /> : <VideoContent alt={profile} src={url} />}
-          </li>
-        ))}
-      </ul>
-      <CommonCard attachment={attachment} />
-    </div>
-  );
+AnimationCardAttachment.defaultProps = {
+  disabled: undefined
 };
 
 AnimationCardAttachment.propTypes = {
@@ -36,7 +21,8 @@ AnimationCardAttachment.propTypes = {
         })
       ).isRequired
     }).isRequired
-  }).isRequired
+  }).isRequired,
+  disabled: PropTypes.bool
 };
 
 export default AnimationCardAttachment;