+++ packages/bundle/src/adaptiveCards/Attachment/OAuthCardAttachment.js
@@ -1,31 +1,14 @@
 import PropTypes from 'prop-types';
-import React, { useMemo } from 'react';
+import React from 'react';
 
-import { hooks } from 'botframework-webchat-component';
+import OAuthCardContent from './OAuthCardContent';
 
-import AdaptiveCardBuilder from './AdaptiveCardBuilder';
-import AdaptiveCardRenderer from './AdaptiveCardRenderer';
-import useAdaptiveCardsPackage from '../hooks/useAdaptiveCardsPackage';
+const OAuthCardAttachment = ({ attachment: { content } = {}, disabled }) => (
+  <OAuthCardContent content={content} disabled={disabled} />
+);
 
-const { useDirection, useStyleOptions } = hooks;
-
-const OAuthCardAttachment = ({ attachment: { content } = {} }) => {
-  const [adaptiveCardsPackage] = useAdaptiveCardsPackage();
-  const [direction] = useDirection();
-  const [styleOptions] = useStyleOptions();
-
-  const builtCard = useMemo(() => {
-    if (content) {
-      const builder = new AdaptiveCardBuilder(adaptiveCardsPackage, styleOptions, direction);
-
-      builder.addCommonHeaders(content);
-      builder.addButtons((content || {}).buttons, true);
-
-      return builder.card;
-    }
-  }, [adaptiveCardsPackage, content, direction, styleOptions]);
-
-  return <AdaptiveCardRenderer adaptiveCard={builtCard} />;
+OAuthCardAttachment.defaultProps = {
+  disabled: undefined
 };
 
 OAuthCardAttachment.propTypes = {
@@ -33,7 +16,8 @@ OAuthCardAttachment.propTypes = {
     content: PropTypes.shape({
       buttons: PropTypes.array
     }).isRequired
-  }).isRequired
+  }).isRequired,
+  disabled: PropTypes.bool
 };
 
 export default OAuthCardAttachment;