+++ packages/bundle/src/adaptiveCards/Attachment/ThumbnailCardAttachment.js
@@ -1,49 +1,14 @@
-/* eslint no-magic-numbers: ["error", { "ignore": [25, 75] }] */
-
 import PropTypes from 'prop-types';
-import React, { useMemo } from 'react';
+import React from 'react';
 
-import { hooks } from 'botframework-webchat-component';
+import ThumbnailCardContent from './ThumbnailCardContent';
 
-import AdaptiveCardBuilder from './AdaptiveCardBuilder';
-import AdaptiveCardRenderer from './AdaptiveCardRenderer';
-import useAdaptiveCardsPackage from '../hooks/useAdaptiveCardsPackage';
+const ThumbnailCardAttachment = ({ attachment: { content }, disabled }) => (
+  <ThumbnailCardContent content={content} disabled={disabled} />
+);
 
-const { useDirection, useStyleOptions } = hooks;
-
-const ThumbnailCardAttachment = ({ attachment: { content } = {} }) => {
-  const [adaptiveCardsPackage] = useAdaptiveCardsPackage();
-  const [direction] = useDirection();
-  const [styleOptions] = useStyleOptions();
-  const builtCard = useMemo(() => {
-    if (content) {
-      const builder = new AdaptiveCardBuilder(adaptiveCardsPackage, styleOptions, direction);
-      const { TextSize, TextWeight } = adaptiveCardsPackage;
-      const { buttons, images, subtitle, text, title } = content;
-      const { richCardWrapTitle } = styleOptions;
-
-      if (images && images.length) {
-        const [firstColumn, lastColumn] = builder.addColumnSet([75, 25]);
-        const [{ tap, url }] = images;
-
-        builder.addTextBlock(
-          title,
-          { size: TextSize.Medium, weight: TextWeight.Bolder, wrap: richCardWrapTitle },
-          firstColumn
-        );
-
-        builder.addTextBlock(subtitle, { isSubtle: true, wrap: richCardWrapTitle }, firstColumn);
-        builder.addImage(url, lastColumn, tap);
-        builder.addTextBlock(text, { wrap: true });
-        builder.addButtons(buttons);
-      } else {
-        builder.addCommon(content);
-      }
-      return builder.card;
-    }
-  }, [adaptiveCardsPackage, direction, content, styleOptions]);
-
-  return <AdaptiveCardRenderer adaptiveCard={builtCard} tapAction={content && content.tap} />;
+ThumbnailCardAttachment.defaultProps = {
+  disabled: undefined
 };
 
 ThumbnailCardAttachment.propTypes = {
@@ -61,7 +26,8 @@ ThumbnailCardAttachment.propTypes = {
       text: PropTypes.string,
       title: PropTypes.string
     }).isRequired
-  }).isRequired
+  }).isRequired,
+  disabled: PropTypes.bool
 };
 
 export default ThumbnailCardAttachment;