+++ packages/bundle/src/adaptiveCards/Attachment/VideoCardAttachment.js
@@ -3,31 +3,14 @@
 import PropTypes from 'prop-types';
 import React from 'react';
 
-import { Components, hooks } from 'botframework-webchat-component';
+import VideoCardContent from './VideoCardContent';
 
-import CommonCard from './CommonCard';
+const VideoCardAttachment = ({ attachment: { content }, disabled }) => (
+  <VideoCardContent content={content} disabled={disabled} />
+);
 
-const { useStyleSet } = hooks;
-const { VideoContent } = Components;
-
-const VideoCardAttachment = ({
-  attachment,
-  attachment: { content: { media, autostart, autoloop, image: { url: imageURL } = {} } = {} } = {}
-}) => {
-  const [{ audioCardAttachment: audioCardAttachmentStyleSet }] = useStyleSet();
-
-  return (
-    <div className={audioCardAttachmentStyleSet}>
-      <ul className="media-list">
-        {media.map(({ url }, index) => (
-          <li key={index}>
-            <VideoContent autoPlay={autostart} loop={autoloop} poster={imageURL} src={url} />
-          </li>
-        ))}
-      </ul>
-      <CommonCard attachment={attachment} />
-    </div>
-  );
+VideoCardAttachment.defaultProps = {
+  disabled: undefined
 };
 
 VideoCardAttachment.propTypes = {
@@ -44,7 +27,8 @@ VideoCardAttachment.propTypes = {
         })
       )
     })
-  }).isRequired
+  }).isRequired,
+  disabled: PropTypes.bool
 };
 
 export default VideoCardAttachment;