+++ packages/testharness/src/host/sendKeys.js
@@ -1,0 +1,10 @@
+import { post } from '../jobs';
+
+export default async function sendKeys(...keys) {
+  return await post({
+    payload: {
+      keys
+    },
+    type: 'send keys'
+  });
+}